refactor: 拆分Stimulus领域实现职责
