+++ server/src/domain/agent/__init__.py
@@ -1,43 +1,44 @@
+from ._stimulus_contract import InvalidStimulusError, StimulusErrorCode
 from .stimulus import (
     ActivityDue,
     ActivityEnded,
-    ActivityFact,
     ActivityObservation,
     ActivityStarted,
-    ActorRef,
-    BodyRegion,
     DailyPlanningDue,
     DeviceConnected,
     DeviceDisconnected,
     DiaryPlanningDue,
-    DynamicMessage,
     DynamicObserved,
     DynamicTargetKind,
-    EvidenceRef,
     ImageMessage,
     ImageSelectionClosed,
     ImageSelectionOpened,
-    InvalidStimulusError,
     InteractionDeadline,
-    MediaRef,
     ProactivePromptDue,
-    ProactiveReason,
-    SongKnowledgeCandidate,
     SongKnowledgeDiscovered,
     SongLearned,
-    SourceRef,
     Stimulus,
-    StimulusErrorCode,
     StimulusKind,
     StimulusSource,
     TextMessage,
-    TouchClickFrequency,
     TouchInteraction,
     ToyVibration,
     UserTyping,
     VoiceMessage,
-    WorldFact,
     WorldObservation,
+)
+from .stimulus_values import (
+    ActivityFact,
+    ActorRef,
+    BodyRegion,
+    DynamicMessage,
+    EvidenceRef,
+    MediaRef,
+    ProactiveReason,
+    SongKnowledgeCandidate,
+    SourceRef,
+    TouchClickFrequency,
+    WorldFact,
     WorldObservationKind,
 )
